# Improve query performance on the certnames table