--- a/ext/test/upgrade-and-exit
+++ b/ext/test/upgrade-and-exit
@@ -83,6 +83,6 @@ psql -U puppetdb puppetdb -c 'select max(version) from schema_migrations;' \
      > "$tmpdir/out"
 cat "$tmpdir/out"
 # This must be updated every time we add a new migration
-grep -qE ' 84$' "$tmpdir/out"
+grep -qE ' 85$' "$tmpdir/out"
 
 test ! -e "$PDBBOX"/var/mq-migrated