--- a/test/puppetlabs/puppetdb/admin_clean_test.clj
+++ b/test/puppetlabs/puppetdb/admin_clean_test.clj
@@ -135,7 +135,7 @@
 (defn purgeable-nodes [node-purge-ttl]
   (let [horizon (time/to-timestamp (time/ago node-purge-ttl))]
     (jdbc/query-to-vec
-     "select * from certnames where deactivated < ? or expired < ?"
+     "select * from certnames_status where deactivated < ? or expired < ?"
      horizon horizon)))
 
 (deftest node-purge-batch-limits